foodgram/requirements.txt
asgiref==3.3.2
Django==3.2.3
gunicorn==20.0.4
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
psycopg2==2.8.6
django-filter==2.3.0
djangorestframework==3.12.4
djoser==2.1.0
reportlab==4.0.9
drf-yasg==1.21.4
